server/requirements.txt
Flask
Flask-Cors
python-dotenv
requests
pymongo
fuzzywuzzy
python-Levenshtein
dnspython
